requirements.txt
Django==3.2
djangorestframework==3.12
psycopg2==2.9
djangorestframework-simplejwt==5.2.2
